Test cases for Roads
Tests for creating roads of different lengths and types.
Testing creating them, making sure they contain Lanes of the right type.
Making sure they are the right width and length.
Testing to see if you can add coordinates for the start and the end of a
road.
Tests for assigning a speedlimit to a road