Commit of updated ListOfListsRoadImpl
Commit of Junction interface
Commit of PreplannedRouteJunction class implementation


The Road now contains a data variable which is the junction at the end
of the road. The access to the junction allows vehicles on this road to
plan their journey over and past the junction node.

Junctions are some collection of LaneSections that connect different
roads to one another.
The produceRoute method allows vehicles to plot their journey accross
the junction. You pass the end coordinate of your current road and the
start coordinate of your destination road.

The PrePlannedRouteJunction is a pre evaluated mapping between roads
connected by the junction.