Tests for the createMappings() method of PrePlannedRouteJunction
Mappings should be generated between entry and exits of junctions.
for example if a junction has one input and one output a mapping should
be created between the input and the output.
for example, if a junction has two inputs and two outputs a mapping
should be created for each one of the inputs to each one of the outputs.

The mappings between input and output are stored in a K,V store. The key
is a string concatenation of the input road coordinates and the output
road coordinates, the value is a List of LaneSection objects that the
Vehicle must follow to go from one road to the start of another.

they should have some naming convention. the coordinateStringFormatter()
method returns a string concatenated in this format->

"inputRoad X, inputRoad Y - outputRoad X, outputRoad Y"